ADD: move system with viewport

+++ client/include/Client.hpp
@@ -30,7 +30,7 @@ class Client {
     std::string _port;
     std::string _ip;
     sf::Event _event;
-    int32_t _viewportX;
+    double _viewportX;
 
   public:
     Client();
@@ -44,8 +44,8 @@ class Client {
     void setUsername(std::string username);
     void setPort(std::string port);
     void setIp(std::string Ip);
-    void setViewport(int32_t viewport);
-    int32_t getViewport();
+    void setViewport(double viewport);
+    double getViewport();
     std::string getUsername();
     std::string getPort();
     std::string getIp();
